+++ Plugins/dateViewer/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.11] - 2021-09-01
+### Added
+- Bold or Italic styles added
+
+### Fixed
+- UTC and 12-hours showing -12
+- Country Language readded
+
 ## [0.2.10] - 2021-08-10
 ### Fixed
 - UTC always showing 00